add "queue". read description for more
I've realized that there is an odd case where .insert() being used twice with large amounts of data will begin to overwrite each other and write to the default location. A queue would likely solve this issue but may also be fixed by making the function asynchronous and adding "await" to db.insert(). I have yet to test this so it's unlikely to do anything, as VS:C states.